dte/requirements.txt
apache-airflow
pandas
requests
psycopg2-binary
